# swagger.yaml
swagger: "2.0"
info:
  description: "Documentação para a API RESTFul desenvolvida no teste backend das Casas Magalhães"
  version: "1.0.0"
  title: "API RESTFul - Teste Backend Casas Magalhães"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "lucas.fber@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:5000"
basePath: "/"
...
schemes:
  - "http"
paths:
...
  /lojas/{id}/produtos:
    post:
      tags:
        - "lojas"
      summary: "Adiciona uma quantidade N de um produto X na loja com o id enviado no path"
      description: ""
      operationId: "addProdutoToLoja"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "Id da loja em que os produtos serão adicionados"
          required: true
          type: "integer"
        - in: "body"
          name: "loja_produto"
          description: "Objeto representando o relacionamento Loja_Produto"
          required: true
          schema:
            $ref: "#/definitions/LojaProduto"
      responses:
        "201":
          description: "Recurso criado com sucesso"
        "404":
          description: "Loja não encontrada"
        "405":
          description: "Entrada inválida"
    get:
      tags:
        - "lojas"
      summary: "Recupera todos os produtos que a loja tem em seu estoque"
      description: ""
      operationId: "GetProdutoByLoja"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "Id da loja que os produtos serão mostrados."
          required: true
          type: "integer"
      responses:
        "200":
          description: "Operação realizada com sucesso"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/LojaProdutoResult"
        "400":
          description: "Id enviado inválido"
        "404":
          description: "Loja não encontrada"
    put:
      tags:
        - "lojas"
      summary: "Atualiza a quantidade de um produtos X numa loja"
      description: ""
      operationId: "updateProdutoQuantidade"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "O id da loja"
          required: true
          type: "integer"
        - in: "body"
          name: "produto_quantidade"
          description: "Objeto representando o produto e a nova quantidade"
          required: true
          schema:
            $ref: "#/definitions/LojaProduto"

      responses:
        "400":
          description: "Id enviado inválido"
        "404":
          description: "Loja não encontrada"
...
definitions:
...
  LojaProduto:
    type: "object"
    required:
      - "produtoId"
      - "quantidade"
    properties:
      produtoId:
        type: "integer"
        example: 2
      quantidade:
        type: "integer"
        example: 10
    xml:
      name: "LojaProduto"
  LojaProdutoResult:
    type: "object"
    properties:
      nome:
        type: "string"
        example: "Smartphone Marca XYZ"
      preco:
        type: "number"
        example: 1289.90
      quantidade:
        type: "integer"
        example: 10
    xml:
      name: "LojaProdutoResult"
...